Update memory configuration to correct RAM size and add detailed SRAM sections (#3)
The code was not executing and crashing immediately as the memory layout
was set incorrectly.
While the STM32H753VIT does have 1MB of RAM, it's not all in one spot
and not all one type.

This change specifies each of the memory regions and how they work.

In a future PR, the RAM tag should be moved to the DTCM section, and a
heap should be configured in AXI, or alternatively, static allocations
for the DMA buffers should be made in AXI/SRAM.
The heap memory should then be used every time that a DMA buffer is
required.

That will make the code faster as the DTCM will be used for all stack
variables.